consolidate timeouts
SensorTag prefers a 10s timeout, so set this and also the
preferred min and max intervals as connection properties.
For 10s timeout a pexpect timeout of 30s (default) is
sufficient.
Half of the timeout is used for the duty cycle.

Teach ble_discovery.py to decode the connection
property characteristic.